Refactor Docker Registry for testing into its own struct
