+++ image/remote_test.go
@@ -1,5 +1,3 @@
-// +build image
-
 package image_test
 
 import (
@@ -33,8 +31,11 @@ func TestRemote(t *testing.T) {
 	t.Parallel()
 	rand.Seed(time.Now().UTC().UnixNano())
 
-	registryPort = h.RunRegistry(t, false)
-	defer h.StopRegistry(t)
+	dockerRegistry := h.NewDockerRegistry()
+	dockerRegistry.Start(t, false)
+	defer dockerRegistry.Stop(t)
+
+	registryPort = dockerRegistry.Port
 
 	spec.Run(t, "remote", testRemote, spec.Sequential(), spec.Report(report.Terminal{}))
 }
